+++ src/main/scala/cognite/spark/v1/Constants.scala
@@ -13,11 +13,11 @@ object Constants {
   val DefaultInferSchemaLimit = 10000
   val DefaultDataPointsLimit = 100000
   val DefaultSequencesLimit = 10000
-  val DefaultPartitions = 1
+  val DefaultPartitions = 200
   val DefaultDataPointsPartitions = 20
-  val DefaultParallelismPerPartition = 1
+  val DefaultParallelismPerPartition = 10
   val DefaultInitialRetryDelay: FiniteDuration = 150.millis
-  val DefaultMaxBackoffDelay: FiniteDuration = 15.seconds
+  val DefaultMaxBackoffDelay: FiniteDuration = 120.seconds
   val DefaultBaseUrl = "https://api.cognitedata.com"
   val SparkDatasourceVersion = s"${BuildInfo.organization}-${BuildInfo.version}"
   val millisSinceEpochIn2100 = 4102448400000L