+++ src/main/scala/cognite/spark/v1/DefaultSource.scala
@@ -161,6 +161,10 @@ class DefaultSource
         new ThreeDModelRevisionNodesRelation(config, modelId, revisionId)(sqlContext)
       case "sequences" =>
         new SequencesRelation(config)(sqlContext)
+      case "labels" =>
+        new LabelsRelation(config)(sqlContext)
+      case "relationships" =>
+        new RelationshipsRelation(config)(sqlContext)
       case _ => sys.error("Unknown resource type: " + resourceType)
     }
   }
@@ -192,8 +196,12 @@ class DefaultSource
           new FilesRelation(config)(sqlContext)
         case "sequences" =>
           new SequencesRelation(config)(sqlContext)
+        case "labels" =>
+          new LabelsRelation(config)(sqlContext)
         case "sequencerows" =>
           createSequenceRows(parameters, config, sqlContext)
+        case "relationships" =>
+          new RelationshipsRelation(config)(sqlContext)
         case _ => sys.error(s"Resource type $resourceType does not support save()")
       }
       val batchSizeDefault = relation match {